common/experiment/experiment.session/pyproject.toml
[tool.poetry]
name = "experiment-session"
version = "0.1.0"
description = ""
authors = ["Damien Bloch <bloc.dam@gmail.com>"]
readme = "README.md"
packages = [{include = "experiment"}]

[tool.poetry.dependencies]
python = "^3.11, <3.12"

asyncpg = "^0.27.0"
platformdirs = "^3.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.4"}

experiment-configuration = { path = "../experiment.configuration", develop = true }
attrs = "^23.1.0"
parameter-types = {path = "../../types/parameter_types", develop = true}
sequence-runtime = {path = "../../sequence/sequence.runtime", develop = true}
sqlalchemy-utils = "^0.41.1"
psycopg2 = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
